% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohazards-package.R
\docType{package}
\name{geohazards-package}
\alias{geohazards}
\alias{geohazards-package}
\title{geohazards: An R package for extracting global climate risk, susceptibility, and vulnerability data}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jorge Ubirajara Pedreira Junior \email{jorge.ubirajara@ufba.br} (\href{https://orcid.org/0000-0002-8243-5395}{ORCID}) [copyright holder]

}
\keyword{internal}
